Added VRTM back

+++ contracts/modules/TransferManager/SingleTradeVolumeRestrictionManager.sol
@@ -64,9 +64,9 @@ contract SingleTradeVolumeRestrictionManager is ITransferManager {
 
         if (isTransferLimitInPercentage) {
             if(specialTransferLimitsInPercentages[_from] > 0) {
-                validTransfer = (_amount.mul(10**18).div(ISecurityToken(securityToken).totalSupply())) <= specialTransferLimitsInPercentages[_from];
+                validTransfer = (_amount.mul(uint256(10)**18).div(ISecurityToken(securityToken).totalSupply())) <= specialTransferLimitsInPercentages[_from];
             } else {
-                validTransfer = (_amount.mul(10**18).div(ISecurityToken(securityToken).totalSupply())) <= globalTransferLimitInPercentage;
+                validTransfer = (_amount.mul(uint256(10)**18).div(ISecurityToken(securityToken).totalSupply())) <= globalTransferLimitInPercentage;
             }
         } else  {
           if (specialTransferLimitsInTokens[_from] > 0) {
